+++ examples/tf/sine_benchmark_fixed_six_break_points_NO_GPU_sp.py
@@ -4,7 +4,7 @@ from time import time
 import os
 os.environ['CUDA_VISIBLE_DEVICES'] = ''
 
-breaks = np.array((0.0, 0.94, 2.96, 4.93, 7.02, 9.04, 10.0))
+breaks = np.array((0.0, 0.94, 2.96, 4.93, 7.02, 9.04, 10.0), dtype=np.float32)
 
 n = np.logspace(3, 7, num=15, dtype=np.int)
 n_repeats = 10
